Extended the Player and Resource classes to the MatterEntity class.
